transport: split underlay and overlay into modules
This will be preferable in the future as underlay functionality expands,
useful to just land this independently so it's not mixed in/to make
future rebases a bit more straightforward.
